refactor: added suggested changes

+++ Lume.Application/Staff/Commands/UpdateStaff/UpdateStaffCommand.cs
@@ -7,6 +7,6 @@ public class UpdateStaffCommand : IRequest<bool>
     public Guid Id { get; set; }
     public string Name { get; set; } = null!;
     public string Surname { get; set; } = null!;
-    public double Salary { get; set; }
+    public int Salary { get; set; }
     public string PhoneNumber { get; set; } = null!;
 }
